packaged for proper usable lib

+++ src/types.zig
@@ -1,2 +1,5 @@
-pub const Unit = usize;
+/// position
+pub const Unit = u16;
+
+/// character type (pixel size)
 pub const Unicode = u21;
